Add files via upload
 This project is a library management system that we made as a group of two. We made it using Java code and object-orientated principles in the BlueJ software. Its main purpose is to manage the books in a library efficiently. This project has 3 classes the book, library and testLibrary class. The book class is used to store the information of individual books, and the library class is aimed at handling the library operations finally the testLibrary class is used to test the functionality of the library/

The book class stores various fields for the details of individual books like the author's name, book title, shelf number or number of pages. These fields are important to help store all the necessary information on each book. We also implemented various accessor and mutator methods more commonly known as getter and setter methods so that we could retrieve information on certain books and set the information like shelf number. At the end of the book class, we have the printIn methods which allow us to display all the information on a certain book

Next, we have the library class. This class allows us to store an array of up to 500 books in a collection. This also includes useful methods such as the addBook method which allows us to add books to the collection as well as the borrowBook and returnBook methods. A very useful method we introduced was the seeAvailableBooks method which allows you to see all books available to borrow in the collection at that time which is helpful for any staff or users to see what they can take. You also have the option to print any general information about the library like its address and opening hours as well as remove a book using its barcode through the methods we added. 

Finally, we used the testLibrary class to test the functionality of our library system. We executed various methods like borrowing and returning books as well as the getter and setter methods. We also tested various functions like the displaying of the library's general info